requirements.txt
pandas>=1.3.0
scikit-learn==1.0.2
numpy>=1.21.0
requests==2.25.1
psycopg2-binary==2.8.6
pytest>=6.0.0
pytest-cov==2.12.1
flask==2.0.1
gunicorn==20.1.0
docker
kubernetes>=18.0.0
mlflow==1.20.0
tensorflow==2.6.0
torch>=1.9.0
torchvision>=0.10.0
boto3>=1.18.0
redis>=3.5.3
celery[redis]>=5.1.2
prometheus-client>=0.11.0
matplotlib>=3.3.0
seaborn>=0.11.0
joblib>=1.0.0
PyYAML>=5.4.0
click>=8.0.0
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.0
sqlalchemy>=1.4.0
alembic>=1.7.0
flake8>=3.9.0
black>=21.0.0
isort>=5.9.0
pre-commit>=2.15.0
